Set inputpath as global variable in main.tmpl
The inputpath (1st argument to main, location of 0.xml file) is exposed as a global variable, such that it can be used by other functions and in different files. E.g. for logging purposes it is convenient to place the log files in the outputpath, which is constructed from the inputpath variable (it is basically set equal to it in the main.c by srtcpy(outputpath, inputpath) ).